services/listening/requirements.txt
PyYAML>=6.0

fastapi>=0.110
uvicorn[standard]>=0.27

# OpenAI STT backend (engine=openai)
openai>=1.0.0
numpy>=1.22
sounddevice>=0.4
